Extract reusable PlotlyChart component
